Split some less used controller out of typescript module refs #736
